[event-input] Enable doubletap emulation by default. Fixes JB#34610
Originally the disable/enable toggle could be used for enabling double
tap detection even in display on states - this is because the feature
was developed primarily with sdk and mouse double clicks in mind.

Later the same detection logic was used for handling double taps that
happen in lpm display states where hw gesture detection is not available.
In these cases the enable/disable setting was in practice ignored.

While fixing the double tap detection during display power off sequence,
I accidentally changed the logic so that it requires
  touch grabbed && setting enabled
instead of the original
  touch grabbed || setting enabled

Since the new behavior actually makes more sense, fix the setting value
so that double tap emulation is enabled by default.