[powerkey] Add state machine for home key handling. JB#34962
On devices that have physical home key, it can used by ui for example
as a shortcut to application grid. However handling home key press
when display is off needs to be subjected to display unblanking policy
which resides in mce.

Add a state machine that deals with input & blanking policies on mce
side and then signals compositor about the home key press when/if
appropriate state is reached and ui should take some action.