[datapipe] Elevate logging of display state requests. JB#36274
Analyzing reasons for pocket calls and other issues related to unwanted
display unblanking is next to impossible if reason for the display wakeup
is not known.

Use a wrapper macro that logs points in mce code base from which display
state request leading to actually powered up states are made.