[display] Allow LPM display states only when lipstick is running. JB#36473
As LPM ui is implemented in lipstick, it makes no sense to allow LPM
display states when lipstick service is not running.

Note: As long as compositor too exists in lipstick process, there is
no need to cancel LPM display states due to lipstick exit as that is
implicitly handled as a result of compositor dropping from D-Bus.