[dbus] Allow use of sender in signal matching. JB#18661
The mce signal listener does not support matching sender names. This is not a
problem as long as the interfaces and member names are unique enough. But when
tracking signals from standard dbus interfaces such as InterfacesAdded and
InterfacesRemoved from org.freedesktop.DBus.ObjectManager / PropertiesChanged
from org.freedesktop.DBus.Properties interface, the sender must be included
in the rule to avoid unnecessary i/o and scheduling.

Make it possible to define sender name when installing mce D-Bus signal
listeners.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>