[display] Start/stop blanking timers based on never-blank mode. JB#37393
Blanking timers are incabable of dimming/powering off the display while
never-blank mode is active. However, leving such timers active and/or
allowing them to be programmed causes side effects such as incorrect
signaling of blanking inhibit state.

Stop blanking timers when never-blank mode is enabled.

Do not allow the timers to be restarted while never-blank mode is active.

Re-evaluate the need for blanking timers when neverblank mode is disabled.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>