[settings] Avoid internal recursion and ensure external notification
When a setting key is changed via D-Bus interface, the value requested
might get filtered by the mce side logic tracking the setting. If the key
gets changed back to the original value, no change notification signal
is emitted and clients can get out of sync with reality.

When a key value changes, make sure a dbus notification for the key is
emitted even if the value gets changed back to the original value.

To avoid internal ringing, do not call internal change notifiers when a
key is set to the same value as it is already holding.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>