[inactivity] Skip activity renew instead of switching to inactivity
Datapipe filter callback functions can only modify the value that datapipe
ends up having and can't be used to skip calling of notification functions.
In case of inactivity processing this means that when filter function does
not allow extending of "user activity" period, it actually immediately
terminates the period.

Instead of having a single inactivity datapipe with filter callback, use two
separate "event" and "state" datapipes chained so that events can be
propagated from event to state datapipe or ignored depending on whether the
activity change is allowed or not.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>