[inactivity] Do not allow activity while lockscreen is active
Conceptually user activity period should not be extended while lockscreen ui
is shown. The rationale for this is: If the display should wake up in a
situation where it should not be allowed (say misbehaving proximity sensor
allows unblanking in pocket), stray input events should not be allowed to
keep the display from blanking.

Drop incorrect "allow activity on tklock is active and display is on" rule.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>