[tklock] Do not allow tkunlock while display is blanked. Fixes MER#1831
If lockscreen blanking timeout happens while user has finished unlock
gesture but has not yet lifted the finger off screen, the gesture gets
finalized due to due to display getting powered off and lockscreen is
left in deactivated state - or in case the lpm feature is enabled, the
display can even bounce back on.

As the intent is that deactivating lockscreen should happen only as a
response to explicit user actions, there should be no need to allow such
actions to be requested when display is off and such interaction with the
user is not really possible.

Enforce this intent and do not allow tkunlock requests made while display
is not in fully powered up state.

Note that this is only protection against accidental unlocking. Applications
can still request display to be powered on. And if that is allowed, they
can also request lockscreen deactivation (which can still be denied due to
other reasons such as device lock being active).

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>