[display] Delay display off signal until backlight is off. Fixes JB#30476
Display off signal is broadcast when ON->OFF transition begins. If the
signal triggers UI side transitions, there can be flickering during
backlight fade-to-black.

Delay sending of display off signal until the backlight fade-to-black
has been finished, so that any changes in display content are invisible
to the user.

Also make it so that querying display state always returns the same value
as what was broadcast earlier.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>