[buttonbacklight] Power off backlight after timeout. JB#42094
The backlight for menu/home/back keys is enabled whenever UI side signals
that key events will be used for UI control. Previously this meant that an
android application is on top. Now that it is possible for these keys to
be used also in SFOS context, it as a side effect means that the backlight
is on practically always.

Keep the logical backlight enable policy as it were, but implement a layer
beneath it to allow powering off the button backlight after 5 seconds of
no significant activity (e.g. display state changes and input events not
related to touch screen).

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>