[buttonbacklight] Add enablers for overriding backlight policy. JB#42094
If menu/home/back keys are used for sailfish applications too, the
button backlight policy interferes with verification test in the CSD
tool.

Add D-Bus methods allowing the CSD application to override button
backlight policy while it is running the verification test.

The required D-Bus constants are defined in mce-dev >= 1.26.0.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>