[powerkey] Do not reset p-sensor override when sensor state is unknown
The emergency proximity sensor override is supposed to works so that
pressing power key 3 times (by default) while sensor is stuck forces the
effective senśor state to OPEN. Resetting the counter while sensor state is
not CLOSED can make it impossible to use the feature for recovering from
malfunctioning on-demand activation getting stuck to UNDEF state.

Reset the counter only when sensor in in OPEN state.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>