[tklock] Keep p-sensor active while handling proximity locking
Proximity locking occurs when proximity sensor is covered after display is
blanked due to inactivity timers. The locking depends on proximity sensor
state, which might not be available in on-demand proximity mode.

To keep proximity sensor powered on, imply need for it while there is
pending proximity locking timeout scheduled.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>