Handle KEY_WAKEUP
The Pine64 Pinephone and PinephonePro have an EG25 modem connected via
USB and GPIO.  When the modem rings, it triggers a GPIO line to the main
CPU to wake it up.  The kernel driver which implements this turns it
into a KEY_WAKEUP key press for each ring.

Use this key event in MCE to temporarily prevent the kernel going back
to sleep while the modem is ringing and keep the screen on.  A timeout
of 10 seconds will allow the kernel to sleep again after KEY_WAKEUP
events stop.