Fix build with Musl libc
RTLD_DEEPBIND is glibc specific and not in POSIX, guard it's use

TEMP_FAILURE_RETRY is glibc specific. If it's not available, define it.
The implementation is taken from Bionic which is permissively licensed,
so should be compatible with the licensing here but IANAL

On 32-bit arches input_event->time is undefined when using a time64
kernel. Instead read the input_event_sec and input_event_usec values
directly and save those instead.
Also make sure to support input_event_sec and input_event_usec on
non-time64 kernels (downstream Android ones for example)