[display] Normalize placeholder topmost window pid values. Fixes JB#54297
It used to be that lipstick reported topmost window pid with value -1
when there was no application windows on screen. However due to some
recent change nowadays (also) pid value zero seems to be used in such
situations. As the pid value is used by mce for detecting situations
where some application window is on screen despite lockscreen being
active, there are subtle problems with for example blanking timers and
fingerprint wakeup handling.

Normalize pid values within mce so that -1 is used whenever lipstick
reports value that is less than or equal to zero.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>