[mce] Musl libc compatibility. JB#49548
Add explanatory comment on why RTLD_DEEPBIND might be needed and
emit a diagnostic message when it is not available.

In order to have input event time stamp macros only in one place, move
them to evdev.h and add an explanatory comment on why they are needed.

While it is not strictly speaking not needed, add customary multiple
include protection to musl-compatibility.h.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>