[audiorouting] Handle "foralien" audio sink suffixes. Fixes JB#50639
When android applications have audio/video calls, proximity blanking is
active regardless of whether headset is used or not. This happens due
to mce not recognizing audiosinks like "NAMEforalien" as aliases for
"NAME" and ends up assuming that earpiece is in use.

Sinks with names like "NAMEforcall" are already treated as aliases for
"NAME" - apply the same logic to sinks with names like "NAMEforalien".

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>