[charging] Apply thresholds after full just once. JB#58372
Make the apply-thresholds-after-full option effective just once - change
to apply-thresholds once fully charged. Do the same also when the
charger gets disconnected, irrespective of the battery level.

Clear battery full seen when changing the mode or it would require to
reconnect the charger in order to set apply-thresholds-after-full after
reaching the upper threshold.

Signed-off-by: Martin Kampas <martin.kampas@jolla.com>