[mcetool] Make it easier to undo ad-hoc settings tweaks. Fixes JB#60163
While mcetool already makes it possible to reset all / some / select
settings back to configured defaults, this requires that user either
knows what the default is (and uses --set-xxx=value option) or what
the relevant setting path is (and uses --reset-settings=path option).

Augment --set-xxx=<value> handling so that "/key" value can be used
for querying the affected setting path and "/reset" can be used for
resetting the value back to configured default.

For the sake of clarity, make --set-xxx options where this is not
applicable print out error message if "/key" or "/reset" is used.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>