[mcetool] Relax boundaries for led breathing battery limit. JB#62736
When led breathing must be done by sw means, is allowed only when
battery is sufficiently full. The battery level limit can be modified
with mcetool.

In devices that do not have a battery mce internally tracks battery
level as -1 / UNKNOWN and mcetool does not allow negative limit to
be set. Which then means that software led breathing can't be enabled.

Relax mcetool sanity checking so that battery level of -1 can be used.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>