[display] Use wakeup sensor to trigger low power mode display. JB#62083
In some devices proximity sensor starts misbehaving if it is kept
powered on for extensive periods of time. As a workaround, these devices
can be configured to use on-demand proximity sensor mode. However this
then effectively breaks low power display mode.

Make it possible to use wakeup sensor instead of / in addition to
proximity sensor for triggering low power mode display

Enable wakeup sensor when display is off and there are no active calls
or alarms.

Actions taken when wakeup sensor event is received depend on proximity
sensor state. If proximity sensor is known to be in not-covered state,
LPM_ON display state can be activated immediately. Otherwise LPM_OFF
display state is activated to effectively wait a while for proximity
sensor up and/or report not-covered state.

As LPM display is opt-in, also this feature is disabled by default
and needs to be enabled from command line

  mcetool --set-wakeup-sensor-mode=enabled

or via corresponding mce configuration file entry

   /system/osso/dsm/display/wakeup_sensor_enabled=true

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>