[mce] Fix potential null arg to strncmp() function. Fixes JB#63136
There is clang warning about code path that can lead to check_rules()
calling strncmp() with a null argument. In addition to that the code
is a bit hard to follow and has other minor issues too.

Refactor code to fix the issues and rename the function to align with
naming convention used in the module.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>